py-autogui/requirements.txt
mcp>=0.1.0
pyautogui>=0.9.54
pillow>=10.0.0